feat: move services folder to out of app folder
