+++ src/app/actions/users/add-user-action.ts
@@ -2,7 +2,7 @@
 
 import { revalidatePath } from 'next/cache'
 
-import { createUser } from '@/app/services/api-users'
+import { createUser } from '@/services/api-users'
 import { getErrorMessage } from '@/utils/getErrorMessage'
 
 export const addUserServerAction = async (formData: FormData) => {
