+++ src/app/auth.ts
@@ -2,7 +2,7 @@ import { z } from 'zod'
 import NextAuth from 'next-auth'
 
 import { authConfig } from '@/app/auth.config'
-import { apiLogin } from '@/app/services/api-login'
+import { apiLogin } from '@/services/api-login'
 import Credentials from 'next-auth/providers/credentials'
 
 export const { auth, signIn, signOut } = NextAuth({
