+++ src/services/api-users.ts
@@ -1,5 +1,5 @@
 import { User } from '@/types'
-import { api } from '@/app/services/api'
+import { api } from '@/services/api'
 
 export async function getUser(id: string) {
   const { data } = await api.get(`/users/${id}`)
